Extract and export the `option` serializer
